pyproject.toml
[tool.poetry]
name = "hello_world"
version = "0.1.0"
description = "A CrewAI project demonstrating multi-model usage with OpenRouter"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = {extras = ["tools"], version = "^0.1.0"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
